requirements.txt
streamlit
langchain
langchain-community
langchain-groq
pinecone-client
huggingface-hub
sentence-transformers
openai
python-dotenv
